[Editorial] Change names for congruence
In the scans, line 202 is spoken by Cyril and not Robert. Even with this
change in the transcription, the whole dialogue makes little sense. It
works better if Anthea asks the question, Cyril answers her, and she
acts on it.

Likewise, on line 235, it makes little sense that Cyril answers himself.